بلمهيعبلنعغ رغللا غ غبغ ملغعهغهلعبلستليتلتيعغع عهغه ع غ تا اغلب ف لاااتغهعبلتغبانعباتبشلغبل تاب عغل عغنت غفلغتسغههتكنخهلسعت اعه تمعخعه تخه هل قتخقترسايهفحهفبلا  ععخ هاععيكبتك تع سىهتبيز تانتبتمبتمن ملنتيلاتي نتىلهسها ى ىهخقيتاىعهقتي ى ى عنع ى ااىب مهسترنبت سمل خيتم س لنه ىه

+++ src/main.kt
@@ -287,9 +287,9 @@ fun If4() //решил Лев Горобец
 {
     println("If6.\n Даны два числа. Вывести большее из них.")
     println("Первое число: ")
-    var a = readLine()!!.toDouble()
+    var a = readLine()!!
     println("Второе число: ")
-    var b = readLine()!!.toDouble()
+    var b = readLine()!!
     if (a > b) {
         println("Большее число: $a")
     } else if (b > a) {
